chore: fix test
