+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/snap-solana-wallet.git"
   },
   "source": {
-    "shasum": "uiNjP7mfyaqqu8ul/tiP+PgiyAbb0gvCRuzydiAyjg4=",
+    "shasum": "2prEQYnglJ1kDkR2DIMlnSDPYluuCKz431uCG9vDJYs=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
