+++ packages/snap/src/core/services/wallet/WalletService.ts
@@ -25,7 +25,6 @@ import { ScheduleBackgroundEventMethod } from '../../handlers/onCronjob/backgrou
 import type { SolanaKeyringAccount } from '../../handlers/onKeyringRequest/Keyring';
 import { addressToCaip10 } from '../../utils/addressToCaip10';
 import { deriveSolanaPrivateKey } from '../../utils/deriveSolanaPrivateKey';
-import { formatSignInMessage } from '../../utils/formatSignInMessage';
 import { getSolanaExplorerUrl } from '../../utils/getSolanaExplorerUrl';
 import type { ILogger } from '../../utils/logger';
 import logger from '../../utils/logger';
@@ -40,6 +39,7 @@ import type { TransactionHelper } from '../execution/TransactionHelper';
 import { mapRpcTransaction } from '../transactions/utils/mapRpcTransaction';
 import type {
   SolanaSignAndSendTransactionResponse,
+  SolanaSignInRequest,
   SolanaWalletRequest,
 } from './structs';
 import {
@@ -367,7 +367,7 @@ export class WalletService {
     const { address } = account;
     const { params } = request.request;
 
-    const messageUtf8 = formatSignInMessage(params);
+    const messageUtf8 = this.#formatSignInMessage(params);
     const messageBytes = getUtf8Codec().encode(messageUtf8);
     const messageBase64 = getBase64Codec().decode(messageBytes);
 
@@ -439,5 +439,76 @@ export class WalletService {
     );
 
     return verified;
+  }
+
+  /**
+   * Formats a Solana Sign-In message into a string.
+   *
+   * @param signInParams - The sign-in message parameters.
+   * @returns The formatted message as a string.
+   */
+  #formatSignInMessage(signInParams: SolanaSignInRequest['params']): string {
+    // ${domain} wants you to sign in with your Solana account:
+    // ${address}
+    //
+    // ${statement}
+    //
+    // URI: ${uri}
+    // Version: ${version}
+    // Chain ID: ${chain}
+    // Nonce: ${nonce}
+    // Issued At: ${issued-at}
+    // Expiration Time: ${expiration-time}
+    // Not Before: ${not-before}
+    // Request ID: ${request-id}
+    // Resources:
+    // - ${resources[0]}
+    // - ${resources[1]}
+    // ...
+    // - ${resources[n]}
+
+    let message = `${signInParams.domain} wants you to sign in with your Solana account:\n`;
+    message += `${signInParams.address}`;
+
+    if (signInParams.statement) {
+      message += `\n\n${signInParams.statement}`;
+    }
+
+    const fields: string[] = [];
+    if (signInParams.uri) {
+      fields.push(`URI: ${signInParams.uri}`);
+    }
+    if (signInParams.version) {
+      fields.push(`Version: ${signInParams.version}`);
+    }
+    if (signInParams.chainId) {
+      fields.push(`Chain ID: ${signInParams.chainId}`);
+    }
+    if (signInParams.nonce) {
+      fields.push(`Nonce: ${signInParams.nonce}`);
+    }
+    if (signInParams.issuedAt) {
+      fields.push(`Issued At: ${signInParams.issuedAt}`);
+    }
+    if (signInParams.expirationTime) {
+      fields.push(`Expiration Time: ${signInParams.expirationTime}`);
+    }
+    if (signInParams.notBefore) {
+      fields.push(`Not Before: ${signInParams.notBefore}`);
+    }
+    if (signInParams.requestId) {
+      fields.push(`Request ID: ${signInParams.requestId}`);
+    }
+    if (signInParams.resources) {
+      fields.push(`Resources:`);
+      for (const resource of signInParams.resources) {
+        fields.push(`- ${resource}`);
+      }
+    }
+    if (fields.length) {
+      message += `\n\n${fields.join('\n')}`;
+    }
+
+    return message;
   }
 }
