+++ packages/snap/src/core/services/wallet/mocks.ts
@@ -129,9 +129,9 @@ export const MOCK_SIGN_IN_RESPONSE: SolanaSignInResponse = {
     address: MOCK_SOLANA_KEYRING_ACCOUNT_2.address,
   },
   signature:
-    '3SAFcibM7NyHVdMNkfXT5BDbYMokLNqSDUyJ1PGWKiFgB7jsRYFLfcvZTjpJ6rLRMog5FtC87D7fX8yGKb66oZwi', // When signed by MOCK_SOLANA_KEYRING_ACCOUNT_2
+    '3SNumB7EXxEw8DbEXjYDKYMZRdGQb8CCrUL5cheTtkHk3arZA7mpvPw7PRT8vYMHzkmcPfrmdar1hofvFmb42rxS', // When signed by MOCK_SOLANA_KEYRING_ACCOUNT_2
   signedMessage:
-    'eyJkb21haW4iOiJleGFtcGxlLmNvbSIsImFkZHJlc3MiOiIyN2g2Y202UzlhZzV5NEFTaTFhMXZiVFNLRXNRTWpFZGZ2WjZhdFBqbWJ1RCIsInN0YXRlbWVudCI6IkkgYWNjZXB0IHRoZSB0ZXJtcyBvZiBzZXJ2aWNlIiwidXJpIjoiaHR0cHM6Ly9leGFtcGxlLmNvbS9sb2dpbiIsInZlcnNpb24iOiIxIiwiY2hhaW5JZCI6InNvbGFuYToxMDEiLCJub25jZSI6IjMyODkxNzU2IiwiaXNzdWVkQXQiOiIyMDI0LTAxLTAxVDAwOjAwOjAwLjAwMFoiLCJleHBpcmF0aW9uVGltZSI6IjIwMjQtMDEtMDJUMDA6MDA6MDAuMDAwWiIsIm5vdEJlZm9yZSI6IjIwMjMtMTItMzFUMDA6MDA6MDAuMDAwWiIsInJlcXVlc3RJZCI6IjEyMyIsInJlc291cmNlcyI6WyJpcGZzOi8vYmFmeWJlaWVteGY1YWJqd2piaWtvejRtYzNhM2RsYTZ1YWwzanNncGRyNGNqcjNvejNldmZ5YXZod3EvIiwiaHR0cHM6Ly9leGFtcGxlLmNvbS9teS13ZWIyLWNsYWltLmpzb24iXX0=', // MOCK_SIGN_IN_REQUEST.params that was JSON.stringified, then base64 encoded
+    'ZXhhbXBsZS5jb20gd2FudHMgeW91IHRvIHNpZ24gaW4gd2l0aCB5b3VyIFNvbGFuYSBhY2NvdW50OgoyN2g2Y202UzlhZzV5NEFTaTFhMXZiVFNLRXNRTWpFZGZ2WjZhdFBqbWJ1RAoKSSBhY2NlcHQgdGhlIHRlcm1zIG9mIHNlcnZpY2UKClVSSTogaHR0cHM6Ly9leGFtcGxlLmNvbS9sb2dpbgpWZXJzaW9uOiAxCkNoYWluIElEOiBzb2xhbmE6MTAxCk5vbmNlOiAzMjg5MTc1NgpJc3N1ZWQgQXQ6IDIwMjQtMDEtMDFUMDA6MDA6MDAuMDAwWgpFeHBpcmF0aW9uIFRpbWU6IDIwMjQtMDEtMDJUMDA6MDA6MDAuMDAwWgpOb3QgQmVmb3JlOiAyMDIzLTEyLTMxVDAwOjAwOjAwLjAwMFoKUmVxdWVzdCBJRDogMTIzClJlc291cmNlczoKLSBpcGZzOi8vYmFmeWJlaWVteGY1YWJqd2piaWtvejRtYzNhM2RsYTZ1YWwzanNncGRyNGNqcjNvejNldmZ5YXZod3EvCi0gaHR0cHM6Ly9leGFtcGxlLmNvbS9teS13ZWIyLWNsYWltLmpzb24=', // MOCK_SIGN_IN_REQUEST.params that was formatted, then base64 encoded
   signatureType: 'ed25519',
 } as const;
